Share get_na_member_lookup between commands
