+++ pombola/south_africa/management/helpers.py
@@ -1,7 +1,11 @@
+from collections import defaultdict
+from itertools import chain
+import re
 import requests
 import time
 import urllib
 
+from pombola.core.models import Position
 
 def fix_province_name(province_name):
     if province_name == 'Kwa-Zulu Natal':
@@ -90,3 +94,55 @@ def geocode(address_string, geocode_cache=None, verbose=True):
         lon = float(geometry['location']['lng'])
         lat = float(geometry['location']['lat'])
         return lon, lat, geocode_cache
+
+title_slugs = ('provincial-legislature-member',
+               'committee-member',
+               'alternate-member')
+
+def get_na_member_lookup():
+    # Build an list of tuples of (mangled_mp_name, person_object) for each
+    # member of the National Assembly and delegate of the National Coucil
+    # of Provinces:
+
+    na_member_lookup = defaultdict(set)
+
+    def warn_duplicate_name(name_form, person):
+        try:
+            message = "Tried to add '%s' => %s, but there were already '%s' => %s" % (
+                name_form, person, name_form, na_member_lookup[name_form])
+            print message
+        except UnicodeDecodeError:
+            print 'Duplicate name issue'
+
+    people_done = set()
+    for position in chain(Position.objects.filter(title__slug='member',
+                                                  organisation__slug='national-assembly'),
+                          Position.objects.filter(title__slug='member-of-the-provincial-legislature').currently_active(),
+                          Position.objects.filter(title__slug='member',
+                                                  organisation__kind__slug='provincial-legislature').currently_active(),
+                          Position.objects.filter(title__slug__in=title_slugs).currently_active(),
+                          Position.objects.filter(title__slug__startswith='minister').currently_active(),
+                          Position.objects.filter(title__slug='delegate',
+                                                  organisation__slug='ncop').currently_active()):
+
+        person = position.person
+        if person in people_done:
+            continue
+        else:
+            people_done.add(person)
+        for name in person.all_names_set():
+            name = name.lower().strip()
+            # Always leave the last name, but generate all combinations of initials
+            name_forms = set(chain(all_initial_forms(name),
+                                   all_initial_forms(name, squash_initials=True)))
+            # If it looks as if there are three full names, try just
+            # taking the first and last names:
+            m = re.search(r'^(\S{4,})\s+\S.*\s+(\S{4,})$', name)
+            if m:
+                name_forms.add(u"{0} {1}".format(*m.groups()))
+            for name_form in name_forms:
+                if name_form in na_member_lookup:
+                    warn_duplicate_name(name_form, person)
+                na_member_lookup[name_form].add(person)
+
+    return na_member_lookup
